Beginning to edit exercise descriptions
My tool checks the last-modified time on each file to determine if it has been edited, so even if I don't need to make changes I need to save it in order to change the timestamp. So you will see some files that only have a newline added at the end of the file -- those are ones that I've checked and don't need editing.